Backend Update
-Added login/sign up functionalities :

The register form now includes a name field.

After registration or login, the user's name is stored in localStorage.

The Navbar displays the user's name along with the Logout button.

Check Authentication State:

Use localStorage.getItem("token") to check if the user is logged in.

Store the result in the isLoggedIn variable.

Logout Function:

Remove the token from localStorage using localStorage.removeItem("token").

Redirect the user to the home page using navigate("/").

Refresh the page with window.location.reload() to update the UI.

Conditional Rendering:

If isLoggedIn is true, show the Logout button.

If isLoggedIn is false, show the Login/Register button.